refactor(providers): move schema normalization out of mod.rs
